refactor(platform): move tool-edit-approval types to their real owner
ToolEditApprovalRequest/Result were defined in src/platform/interfaces.ts,
labeled "Platform-level contract" right next to a comment explicitly saying
they are *not* wired through the Platform object (hosts get them via
SessionHandle.interactions instead). That file's own module doc says it
holds only "the host-neutral interfaces a host wires into initPlatform()",
so the types didn't belong there, and src/tools/approval/toolEditApproval.ts
— which owns every function that builds and consumes them — had to import
them back from Platform and re-export. Two other production files reached
past that re-export straight into @platform/interfaces for the same types,
so callers had three different import paths for one pair of types.

Define the types where the behavior lives (toolEditApproval.ts) and point
the two direct importers (HostInteractions.ts, progressHostInteractions.ts)
and one test file at that source instead of @platform/interfaces. No
behavior change; @tools/approval's existing barrel re-export is unaffected.
