+++ src/test-kernel/desktop/DesktopToolEditApproval.vitest.mts
@@ -12,15 +12,15 @@ import type { SessionEvent } from '@agent/runtime/SessionEventHub';
 import type { DesktopAgentExecutionHost } from '@desktop/main/desktopAgentExecutionHost';
 import type { DiffOptions, DiffSession, DiffSource } from '@hosts/uiHosts';
 
-import type {
-  ToolEditApprovalRequest,
-  ToolEditApprovalResult,
-} from '@platform/interfaces';
 import type { ToolEditPermission } from '@shared/schemas';
 import { SESSION_DISPOSED_CAUSE } from '@shared/copy/interactionCancellation';
 import type { ToolEditApprovalAction } from '@shared/schemas/prompts';
 import { createModuleMocks } from '@test/support/moduleMocks';
 import { createTestSession } from '@test/support/sessionTestUtils';
+import type {
+  ToolEditApprovalRequest,
+  ToolEditApprovalResult,
+} from '@tools/approval/toolEditApproval';
 import { delay } from '@utils/core';
 import {
   createStubDesktopAgentExecutionHost,
